+++ src/ReadGui.java
@@ -1,11 +1,11 @@
 public class ReadGui {
-    public int[] getGuiVector() {
+    public double[] getGuiVector() {
         return guiVector;
     }
 
-    private int[] guiVector;
+    private double[] guiVector;
     public ReadGui(String text){
-        guiVector = new int[26];
+        guiVector = new double[26];
         text= text.toUpperCase();
         for(int i=0;i<text.length();i++){
             int charASCII = (int)text.charAt(i)-65;
@@ -16,7 +16,7 @@ public class ReadGui {
 
 
     }
-    public void showVector(int[] vector){
+    public void showVector(double[] vector){
         for (int i=0; i<vector.length;i++){
             System.out.println(i+"  "+vector[i]);
         }
